backend/requirements.txt
Django==2.0.2
djangorestframework==3.7.7
simplejson==3.13.2
djoser
django-cors-headers
djangorestframework-jwt